Moving command_builder to builder module
