+++ cmd/kpng/builder/command_builder.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package storecmds
+package builder
 
 import (
 	"context"
